+++ CHANGELOG.rst
@@ -524,7 +524,7 @@ Added
 - Sensitivity analysis: an abstract class with sorting, plotting and comparison methods,
   with a dedicated factory and new features (correlation coefficients and Morris indices).
 - Sensitivity analysis: examples.
-- ConcatenationDiscipline: a new discipline to concatenate inputs variables into a single one.
+- Concatenater: a new discipline to concatenate inputs variables into a single one.
 - Gantt chart generation to visualize the disciplines execution time.
 - An interactive web N2 chart allowing to expand or collapse the groups of strongly coupled disciplines.
 - Support pSeven algorithms for single-objective optimization.